pyproject.toml
[project]
name = "lisa-acv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alpha-clip",
    "bitsandbytes>=0.44.1",
    "clip",
    "ftfy>=6.3.1",
    "gradio>=5.8.0",
    "ipykernel>=6.29.5",
    "loralib>=0.1.2",
    "matplotlib>=3.9.2",
    "numpy>=2.1.3",
    "opencv-python>=4.10.0.84",
    "peft>=0.13.2",
    "pillow>=11.0.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "regex>=2024.11.6",
    "scikit-learn>=1.6.0",
    "scipy>=1.14.1",
    "segment-anything",
    "shapely>=2.0.6",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "tqdm>=4.67.1",
    "transformers==4.46.0",
    "wandb>=0.19.0",
]

[tool.uv.sources]
segment-anything = { git = "https://github.com/facebookresearch/segment-anything.git" }
alpha-clip = { path = "../../models/AlphaCLIP" }
clip = { git = "https://github.com/openai/CLIP.git" }
